Rename PDF modal to FileButton
