+++ src/routes/(app)/documents/FileButton.svelte
