+++ src/routes/(app)/documents/requirements/Folder.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import type { SuperValidated } from "sveltekit-superforms";
   import DeleteFileForm from "../DeleteFileForm.svelte";
-  import File from "../File.svelte";
+  import FileLink from "../FileLink.svelte";
   import type { DeleteSchema } from "./+page.server";
   import { isAuthorized } from "$lib/utils/authorization";
   import apiNames from "$lib/utils/apiNames";
@@ -44,7 +44,7 @@
         />
       {:else}
         <div class="flex flex-row gap-1">
-          <File name={folder.name} url={folder.url} />
+          <FileLink name={folder.name} url={folder.url} />
           {#if isAuthorized(apiNames.FILES.BUCKET(PUBLIC_BUCKETS_DOCUMENTS).DELETE, $page.data.user) && isEditing}
             <DeleteFileForm
               fileId={folder.id}
